+++ modules/Food_Service/Students/Transactions.php
@@ -109,7 +109,8 @@ if ( UserStudentID()
 		AND (fst.STUDENT_ID IS NULL OR fst.STUDENT_ID='" . UserStudentID() . "')
 		AND fst.TIMESTAMP BETWEEN CURRENT_DATE
 		AND (CURRENT_DATE + INTERVAL " . ( $DatabaseType === 'mysql' ? '1 DAY' : "'1 DAY'" ) . ")
-		AND fsti.TRANSACTION_ID=fst.TRANSACTION_ID" );
+		AND fsti.TRANSACTION_ID=fst.TRANSACTION_ID
+		ORDER BY fst.TRANSACTION_ID,fsti.ITEM_ID" );
 
 		// TODO: code duplication!
 		/**