+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 10.8.4
 -----------------
 - Fix SQL error escape menu title in DailyMenus.php, ServeMenus.php
 - Fix MySQL date interval in Food_Service/Students/ServeMenus.php
+- SQL order Transactions by ID in Transactions.php
 
 Changes in 10.8.3
 -----------------